% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ordinal_priors.R
\name{get_ordinal_priors}
\alias{get_ordinal_priors}
\title{Get priors for an ordinal variable converted into cumulative log odds}
\usage{
get_ordinal_priors(probs)
}
\arguments{
\item{probs}{A vector of probabilities for each threshold (i.e response point) on your scale.
Probabilities must sum to 1 or to 100.}
}
\value{
A vector of cumulative log odds for all n-1 thresholds (i.e. response points).
  The cumulative log odds for the nth threshold are not returned, given that :
  1) the log odds for the last threshold does not need to be specified as a prior in bmrs;
  2) probabilities nessarily sum to 1 and the log odds for 1 is Inf.
}
\description{
Get priors for an ordinal variable converted into cumulative log odds
}
