% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_qualtrics.R
\name{read_qualtrics}
\alias{read_qualtrics}
\title{Use read_csv on files downloaded from Qualtrics}
\usage{
read_qualtrics(file)
}
\arguments{
\item{file}{A path to a csv file imported from Qualtrics}
}
\value{
A tibble().
}
\description{
Data files exported from Qualtrics are famously difficult to import in R as the
first three rows contain column names, text from the prompts and question labels.
The second row contains the most user friendly information, but because it is not possible
to skip both the first and the third row when reading a file, this usually leads to the
annoying consequence that your first row contains strings, which are either converted to factors
 or force you to specify the type for all columns.
 Use \code{read_qualtrics} instead to directly import your csv file with usable column names and
 allow \code{read_csv} to impute your column types.
}
\examples{
\dontrun{
read_qualtrics("qualtrics_import.csv")
}

}
